launchpad: Refactor fee table generation
