+++ packages/launchpad/types/gas.d.ts
@@ -1,16 +1,14 @@
 import { StdFee } from "./types";
-/**
- * These fees are used by the higher level methods of SigningCosmosClient
- */
-export interface FeeTable {
-  readonly send: StdFee;
-}
 export declare class GasPrice {
   readonly amount: number;
   readonly denom: string;
   constructor(amount: number, denom: string);
 }
-export declare type GasLimits = {
-  readonly [key in keyof FeeTable]: number;
+export declare type GasLimits<T extends Record<string, StdFee>> = {
+  readonly [key in keyof T]: number;
 };
-export declare function buildFeeTable({ denom, amount }: GasPrice, gasLimits: GasLimits): FeeTable;
+export declare function buildFeeTable<T extends Record<string, StdFee>>(
+  gasPrice: GasPrice,
+  defaultGasLimits: GasLimits<T>,
+  gasLimits: Partial<GasLimits<T>>,
+): T;
